Conflicts:
	src/OS_linux/my_gui_plotterClock.cpp

+++ src/1_reader/fake_reader.cpp
@@ -79,7 +79,7 @@ bool readLineFromQueue(Queue &queue, char *pBuffer , uint32_t & sizeOfBuffer)
 			queue.push(line[i]);
 		}
 
-		for (int i = 0; i < numberOfItems; i++)
+		for (size_t i = 0; i < numberOfItems; i++)
 		{
 			char byte;
 			queue.pop(byte);
@@ -92,6 +92,8 @@ bool readLineFromQueue(Queue &queue, char *pBuffer , uint32_t & sizeOfBuffer)
 
 	return retval;
 }
+
+#include <unistd.h>
 
 void serveClient()
 {
@@ -133,9 +135,6 @@ void serveClient()
 
 		LOG("numberOfPushes: " << queue.numberOfPushes);
 		LOG("numberOfPops: " << queue.numberOfPops);
-
-		//parser_update(line);
-
 	}
 }
